requirements.txt
selenium
beautifulsoup4
google-cloud-storage
google-cloud-bigquery
python-dotenv